Added neural interface classes.
Each agent has an object of type NInterface that stores data representing
the agent's afferent and efferent organs. Afferent sensor data is written
to by the simulation (to update the state of each sensor) and can be read
by a neural network. A neural network should write to the efferent sensor
data fields (updating motor output), which can be read by the simulation
for updating the physical representation of the agent in the world.