Added NeatNode class to Genome class.
Each genome now has an array list of NeatNode objects. When a genome is
constructed, a collection of NeatNode objects must also be provided
along with the edge genes - this new way of instanciating Genome objects
was incorporated into the Environment class.

NeatNodes also have a factory method for creating NeatNode objects that
encapsulate RS-neuron MNeuronParams (this helps greatly when creating
genomes for the first time and when adding new nodes via mutation for
example).

TODO: Mutations and other methods in God require updating to use Genomes
with this addition (Ruth - help!). :)